+++ indico/modules/events/abstracts/forms.py
@@ -19,8 +19,9 @@ from indico.core.db import db
 from indico.core.db.sqlalchemy.descriptions import RenderMode
 from indico.modules.events.abstracts.fields import (AbstractField, AbstractPersonLinkListField, EmailRuleListField,
                                                     TrackRoleField)
-from indico.modules.events.abstracts.models.abstracts import EditTrackMode
+from indico.modules.events.abstracts.models.abstracts import AbstractState, EditTrackMode
 from indico.modules.events.abstracts.models.reviews import AbstractAction, AbstractCommentVisibility
+from indico.modules.events.abstracts.placeholders import AbstractInvitationURLPlaceholder
 from indico.modules.events.abstracts.settings import (AllowEditingType, BOACorrespondingAuthorType, BOALinkFormat,
                                                       BOASortField, SubmissionRightsType, abstracts_settings)
 from indico.modules.events.contributions.models.types import ContributionType
@@ -428,8 +429,9 @@ class EditEmailTemplateRuleForm(IndicoForm):
     title = StringField(_('Title'), [DataRequired()])
     rules = EmailRuleListField(_('Rules'), [DataRequired()])
 
-    def __init__(self, *args, **kwargs):
-        self.event = kwargs.pop('event')
+    def __init__(self, *args, event, **kwargs):
+        self.event = event
+        self.email_tpl = kwargs['obj']
         super().__init__(*args, **kwargs)
         self.rules.event = self.event
 
@@ -450,13 +452,30 @@ class EditEmailTemplateTextForm(IndicoForm):
     subject = StringField(_('Subject'), [DataRequired()])
     body = TextAreaField(_('Body'), [DataRequired()])
 
-    def __init__(self, *args, **kwargs):
-        self.event = kwargs.pop('event')
+    def __init__(self, *args, event, **kwargs):
+        self.event = event
+        self.email_tpl = kwargs['obj']
         super().__init__(*args, **kwargs)
         choices = [('', config.NO_REPLY_EMAIL)]
         choices += list(self.event.get_allowed_sender_emails(extra=self.reply_to_address.object_data).items())
         self.reply_to_address.choices = choices
         self.body.description = render_placeholder_info('abstract-notification-email', event=self.event)
+
+    def validate_body(self, field):
+        # disallow using the invitation_url placeholder for non-invited email templates and vice versa
+        invited_rule_present = any(
+            AbstractState.invited.value in value
+            for rule in self.email_tpl.rules
+            for value in rule.values()
+        )
+        has_invitation_link_placeholder = AbstractInvitationURLPlaceholder.is_in(field.data, abstract=None)
+
+        if invited_rule_present and not has_invitation_link_placeholder:
+            raise ValidationError(_('Invitation email templates must contain the {placeholder} placeholder')
+                                  .format(placeholder='{invitation_url}'))
+        elif not invited_rule_present and has_invitation_link_placeholder:
+            raise ValidationError(_('Only invitation email templates may contain the {placeholder} placeholder')
+                                  .format(placeholder='{invitation_url}'))
 
 
 class CreateEmailTemplateForm(EditEmailTemplateRuleForm):