+++ CHANGES.rst
@@ -42,6 +42,8 @@ Bugfixes
 - Fix linebreak display in markdown code blocks in survey section descriptions (:pr:`6553`)
 - Include attached pictures when downloading registration attachments (:pr:`6564`)
 - Only allow marking unpaid registrations as paid (:issue:`6330`, :pr:`6578`)
+- Do not allow mixing notification rules for invited abstracts with other rules (:issue:`6563`,
+  :pr:`6567`)
 
 Accessibility
 ^^^^^^^^^^^^^